+++ src/main/java/com/gadator/gadator/service/UserDetailsServiceImpl.java
@@ -20,7 +20,7 @@ import java.util.List;
 
 @Service
 @Transactional
-public class MyUserDetailsService implements UserDetailsService{
+public class UserDetailsServiceImpl implements UserDetailsService{
 
     @Autowired
     private UserRepository userRepository;
